[Akshay] Edit the distribute cards function of the game
What changes have you made?
I have made changes in the callBreak.js file (in distribute prototype function of the Create Game Constructor).
*

Why do we need this change?
We need this change so that we can remove these Errors:-
	- No spade in the hands.
	- No cards with higher rank than 10.
*

What can this change impact?
This change will support our game to provide correct cards to the user and to provide better equality in game.
*