[Akshay] Edited the handCards checking function(isHandsCardsAreCorrect) in pointTable.js file
What changes have you made?
I have splited the cards checking(are correct or not) function in much simpler form.
*

Why do we need this change?
This change will allow us to read code easily and to understand what it is doing.
*

What can this change impact?
This change will make the code simple and also helps in future implemation.
*