Made some changes in throwable function.
What changes have you made?
made some functions based on condition of a card being thrown.
*

Why do we need this change?
we need this change so that player can determine which card should be thrown in differen conditions.
*

What can this change impact?
it make game comfortable for a player.

*