[Akshay]
What changes have you made?
I have refactored the pointTable.js and now pointTable is in a form of constructor with different fields.
*

Why do we need this change?
These changes are necessary to test the logic in a better way and also to store players points in a file.
These changes also necessary to calculate total score of each round.Now pointTable can also tell the winner of
each round and each game.
*

What can this change impact?
These changes help in handling the each round data,still lots of things to change in pointTable.
Most of the part is not tested which can be create some problem in game.
*